split store.rs file
