+++ crates/turbopack-trace-server/src/main.rs
@@ -8,6 +8,8 @@ use self::{reader::TraceReader, server::serve, store_container::StoreContainer};
 mod reader;
 mod server;
 mod span;
+mod span_graph_ref;
+mod span_ref;
 mod store;
 mod store_container;
 mod u64_empty_string;
